Remove more outdated classes and files
