+++ Archs/ARM/Arm.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Archs/Architecture.h"
-#include "Util/CommonClasses.h"
 #include "Pool.h"
 #include "Core/Expression.h"
 #include "Parser/Tokenizer.h"
