+++ Archs/MIPS/MipsElfFile.h
@@ -2,7 +2,6 @@
 #include "Core/ELF/ElfFile.h"
 #include "Core/FileManager.h"
 #include "Commands/CAssemblerCommand.h"
-#include "Util/CommonClasses.h"
 #include "Mips.h"
 
 class MipsElfFile: public AssemblerFile
