+++ Commands/CDirectiveMessage.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "Util/CommonClasses.h"
 #include "Commands/CAssemblerCommand.h"
 #include "Core/Common.h"
 #include "Core/Expression.h"
