+++ Core/Assembler.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "../Util/CommonClasses.h"
 #include "../Util/FileClasses.h"
 #include "../Util/Util.h"
 #include "FileManager.h"
