+++ Main/Tests.cpp
@@ -1,7 +1,6 @@
 #include "stdafx.h"
 #include "Tests.h"
 #include "Util/Util.h"
-#include "Util/CommonClasses.h"
 #include "Core/Common.h"
 #include "Core/Assembler.h"
